requirements.txt
# Azure Functions and Service Bus dependencies (Updated Sept 2025)
azure-functions>=1.23.0,<2.0.0
azure-servicebus>=7.14.0,<8.0.0
grpcio>=1.62.0  # Required for Python 3.12 compatibility

# Core Azure libraries
azure-core>=1.30.0,<2.0.0
azure-identity>=1.17.0,<2.0.0

# Configuration and validation with Pydantic v2
pydantic>=2.9.0,<3.0.0
pydantic-settings>=2.9.0,<3.0.0

# Monitoring and telemetry
azure-monitor-opentelemetry>=1.5.0,<2.0.0

# Standard library typing extensions (for Python 3.10+ compatibility)
typing-extensions>=4.12.0,<5.0.0

# Development and testing dependencies
pytest>=8.3.0,<9.0.0
pytest-mock>=3.14.0,<4.0.0
pytest-asyncio>=0.24.0,<1.0.0
pytest-cov>=5.0.0,<6.0.0

# Code quality tools (for development)
ruff>=0.7.0,<1.0.0
mypy>=1.11.0,<2.0.0

# Security fixes for known vulnerabilities
# setuptools: Fix for CVE-2022-40897, CVE-2025-47273, CVE-2024-6345
setuptools>=78.1.1,<79.0.0
# black: Fix for CVE-2024-21503
black>=24.3.0,<25.0.0

# Security scanning tools
bandit[toml]>=1.7.0,<2.0.0
safety>=3.2.0,<4.0.0
